requirements.txt
# GodTier Shorts - Backend Dependencies
# Reference toolchain: Python 3.13.x

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.9

# Data Validation
pydantic>=2.5.0

# Environment Variables
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Security / Crypto
PyJWT>=2.8.0
cryptography>=42.0.0

# AI / ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
faster-whisper>=1.0.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# LLM API
openai>=1.3.0

# HTTP Client (for yt-dlp)
httpx>=0.25.0

# yt-dlp for YouTube downloading
yt-dlp>=2024.0.0
